+++ service/controller/resource/monitoring/remotewriteapiendpointsecret/doc.go
@@ -1,0 +1,5 @@
+package remotewriteapiendpointsecret
+
+// This packages creates the remote write secret
+// to be used by the remote write agent so it can access
+// the management cluster prometheus through an ingress.