+++ service/controller/resource/monitoring/remotewriteingress/v1beta1/resource.go
@@ -10,6 +10,7 @@ import (
 	"github.com/giantswarm/micrologger"
 	networkingv1beta1 "k8s.io/api/networking/v1beta1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/intstr"
 
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/generic"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/key"
@@ -58,9 +59,10 @@ func getObjectMeta(v interface{}, config Config) (metav1.ObjectMeta, error) {
 	}
 
 	return metav1.ObjectMeta{
-		Name:      fmt.Sprintf("prometheus-%s-remote-write", key.ClusterID(cluster)),
-		Namespace: key.Namespace(cluster),
-		Labels:    key.PrometheusLabels(cluster),
+		Name:        fmt.Sprintf("prometheus-%s-remote-write", key.ClusterID(cluster)),
+		Namespace:   key.Namespace(cluster),
+		Labels:      key.PrometheusLabels(cluster),
+		Annotations: key.RemoteWriteAuthenticationAnnotations(),
 	}, nil
 }
 
@@ -69,7 +71,60 @@ func toIngress(v interface{}, config Config) (metav1.Object, error) {
 		return nil, nil
 	}
 
-	return &networkingv1beta1.Ingress{}, nil
+	objectMeta, err := getObjectMeta(v, config)
+	if err != nil {
+		return nil, microerror.Mask(err)
+	}
+
+	cluster, err := key.ToCluster(v)
+	if err != nil {
+		return nil, microerror.Mask(err)
+	}
+
+	// Note we only configure a path that will cause a location/HTTP path to be
+	// added to the proxy for the base domain here.
+	//
+	// The common server configuration like TLS is configured in the `Ingress`
+	// installed by PMO chart itself. We want to avoid duplicating the TLS
+	// configuration here as only one certificate can be used for a given
+	// domain, and if multiple `Ingress` resources specify that configuration
+	// the Ingress Controller just picks a random one (first one it finds). And
+	// if it picks up a copied certificate, and that copy happens to become out
+	// of date at some point, it would break HTTPS access to that domain.
+	//
+	// So we want TLS configuration to be controlled only by the `Ingress`
+	// resource that also defines the source of the certificates (i.e. the
+	// Let's Encrypt annotation or the static source for the installation)
+	// so we know as soon as it's updated IC will be using it.
+	ingress := &networkingv1beta1.Ingress{
+		TypeMeta: metav1.TypeMeta{
+			APIVersion: networkingv1beta1.SchemeGroupVersion.Version,
+			Kind:       "Ingress",
+		},
+		ObjectMeta: objectMeta,
+		Spec: networkingv1beta1.IngressSpec{
+			Rules: []networkingv1beta1.IngressRule{
+				{
+					Host: config.BaseDomain,
+					IngressRuleValue: networkingv1beta1.IngressRuleValue{
+						HTTP: &networkingv1beta1.HTTPIngressRuleValue{
+							Paths: []networkingv1beta1.HTTPIngressPath{
+								{
+									Path: fmt.Sprintf("/%s/api/v1/write", key.ClusterID(cluster)),
+									Backend: networkingv1beta1.IngressBackend{
+										ServiceName: key.PrometheusServiceName,
+										ServicePort: intstr.FromInt(int(key.PrometheusPort())),
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+	}
+
+	return ingress, nil
 }
 
 func hasChanged(current, desired metav1.Object) bool {